Added getNameForAddr()
Used a more useful definition for the DnsGetCacheDataTable function
pointer (it still works just as well, but doesn't require a malloc). The
function is not called anywhere, but I have tested it (see commented out
printfs in the function itself). Output is not quite what I expected,
but it's the best I can do for now.